third_party/blink/web_tests/external/wpt/css/css-text/text-spacing-trim/text-spacing-trim-combinations-001-expected.txt

This is a testharness.js-based test.
FAIL "国））国" assert_equals: expected "FHFF" but got "FFFF"
FAIL "国）））国" assert_equals: expected "FHHFF" but got "FFFFF"
Harness: the test ran to completion.
